+++ src/fr/maxlego08/menu/api/font/FontImage.java
@@ -1,0 +1,7 @@
+package fr.maxlego08.menu.api.font;
+
+public interface FontImage {
+
+    String replace(String string);
+
+}