+++ src/fr/maxlego08/menu/api/InventoryManager.java
@@ -464,10 +464,25 @@ public interface InventoryManager extends Savable, Listener {
      */
     int getMaxPage(OfflinePlayer player);
 
+    /**
+     * Retrieves a fake inventory for testing or simulation purposes.
+     *
+     * @return An instance of {@link InventoryDefault} representing the fake inventory.
+     */
     InventoryDefault getFakeInventory();
 
+    /**
+     * Provides access to the enchantments management interface.
+     *
+     * @return An instance of {@link Enchantments} for managing custom enchantments.
+     */
     Enchantments getEnchantments();
 
+    /**
+     * Retrieves the font image utility for managing custom fonts.
+     *
+     * @return An instance of {@link FontImage} for handling font images.
+     */
     FontImage getFontImage();
 
 }